Remove version number from package.json and add homepage
I'm not publishing this package, so according to the documentation[0]
version number is an optional field. As I don't really have any
standards for bumping this number (it's only really done when I feel
like it, which is usually only for big rewrites), I think it's neater to
just not include it.

[0] https://docs.npmjs.com/cli/v9/configuring-npm/package-json#version